Update getpicture.js
I am not sure if it will work correctly with the userinput but all the console logs seem to log what i wanted it to. 

User input = true then it will go on and randomize a photo from that breed. But that would be returned as a link to the photo so have to save that link inside an image tag. 

If what userinput was not valid then it will prompt that it was invalid.